renamed common rnn.py -> time_series.py
