+++ lightwood/encoders/__init__.py
@@ -9,7 +9,7 @@ from lightwood.encoders.text.vocab import VocabularyEncoder
 from lightwood.encoders.text.rnn import RnnEncoder as TextRnnEncoder
 from lightwood.encoders.categorical.onehot import OneHotEncoder
 from lightwood.encoders.categorical.autoencoder import CategoricalAutoEncoder
-from lightwood.encoders.time_series.rnn import TimeSeriesEncoder as TsRnnEncoder
+from lightwood.encoders.time_series.time_series import TimeSeriesEncoder as TsRnnEncoder
 # from lightwood.encoders.audio.amplitude_ts import AmplitudeTsEncoder
 from lightwood.encoders.categorical.multihot import MultiHotEncoder
